requirements.txt
munch
numpy
Pillow
tqdm
soundfile
pydub
scipy
pyyaml
librosa
matplotlib
ipykernel
jupyter
parallel_wavegan
click
tensorboard
setuptools==59.5.0
sounddevice==0.4.4